+++ R/discomb.R
@@ -547,6 +547,11 @@ discomb <- function(TE, seTE,
     treat2 <- as.character(treat2)
   if (is.factor(studlab))
     studlab <- as.character(studlab)
+  ##
+  ## Remove leading and trailing whitespace
+  ##
+  treat1 <- rmSpace(rmSpace(treat1, end = TRUE))
+  treat2 <- rmSpace(rmSpace(treat2, end = TRUE))
   
   
   ##